Restructure done, change all links for frontend part and backend part, 3 pkg.json (client, server, root)
