+++ client/src/components/LeafletCenterViewHandler.js
