+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./client/App";
+import App from "./App.js";
 import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(
