+++ package.json
@@ -20,24 +20,13 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "mongoose": "^6.0.12",
-    "leaflet": "^1.7.1",
-    "leaflet.markercluster": "^1.5.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-leaflet": "^3.2.2",
-    "react-leaflet-markercluster": "^3.0.0-rc1",
-    "react-router-dom": "^6.0.1",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.3",
-    "uuid": "^8.3.2",
+    "concurrently": "^6.4.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "server": "cd ./server && npm run server",
+    "client": "cd ./client && npm run start",
+    "work": "concurrently \"npm run server\" \"npm run client\""
   },
   "eslintConfig": {
     "extends": [
