+++ server/controllers/ControllerLogin.js
@@ -1,3 +1,4 @@
+/* eslint-disable consistent-return */
 import UserModel from "../models/UserModel.js";
 import bcrypt from "bcrypt";
 
@@ -10,29 +11,29 @@ const login = async (request, response) => {
 
         // Validation of the user
 
-        if (!email || !password)
+        if (!email || !password) {
             return response
                 .status(400)
                 .json({errorMessage: "Please enter all required fields."});
+        }
 
         const existingUser = await UserModel.findOne({UserEmail: email});
 
-        if (existingUser)
+        if (existingUser) {
             return bcrypt.compare(
                 password,
                 existingUser.UserPasswordHash,
                 (error, result) => {
                     if (result) {
                         return response.json("You are logged in !");
-                    } else {
-                        return response
-                            .status(401)
-                            .json("Wrong email or password :(");
                     }
+                    return response
+                        .status(401)
+                        .json("Wrong email or password :(");
                 },
             );
+        }
     } catch (error) {
-        console.error(error);
         response.status(500).send("Check the terminal");
     }
 };
